feat: working deploy with full code
