+++ backend/functions/package.json
@@ -15,7 +15,15 @@
   "main": "lib/index.js",
   "dependencies": {
     "firebase-admin": "^10.2.0",
-    "firebase-functions": "^3.21.0"
+    "firebase-functions": "^3.21.0",
+    "cookie-parser": "^1.4.5",
+    "axios": "^0.21.1",
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.21",
+    "nanoid": "^3.1.22",
+    "octokit": "^2.0.10"
   },
   "devDependencies": {
     "typescript": "^4.6.4"
