+++ backend/functions/src/util/responseShape.ts
